# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Restaurant App API",
    "version": "1.0.0",
    "description": "API documentation for Restaurant ordering system (Users, Riders, Restaurants, Orders)."
  },
  "servers": [
    {
      "url": "http://localhost:5000/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Order": {
        "type": "object",
        "required": ["restaurant_id", "items"],
        "properties": {
          "restaurant_id": { "type": "integer", "example": 1 },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "menuItemId": { "type": "integer", "example": 2 },
                "quantity": { "type": "integer", "example": 3 }
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/riders/{id}": {
      "get": {
        "summary": "Get rider by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Rider details" }
        }
      },
      "delete": {
        "summary": "Delete a rider (admin only)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Rider deleted" }
        }
      }
    },
...
    "/orders": {
      "get": {
        "summary": "Get all orders (admin) or own orders (customer)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List of orders" }
        }
      },
      "post": {
        "summary": "Create a new order",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Order" }
            }
          }
        },
        "responses": {
          "201": { "description": "Order created successfully" }
        }
      }
    }
  }
}